requirements.txt
Flask==2.3.3
pypdf==3.17.0
Pillow==10.0.0
pytesseract==0.3.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.0
openai-whisper==20231117
numpy==1.24.3